refactor: add some logs and refactor code
